ADD FEED OF OINKS
We worked on the necessary methods and logic to get the User's followers' oinks in order. Due to how our database is set up we find each oink object from a user we follow, add them to an array and then sort that array by the created_at column and then reverse. This is probably not the best approach. Consider optimizing.