FIX CURRENT USER LOADING
The first implementation had us using the session to get the profile which made clicking on other's profiles have your own oinks. The same thing was happening with the gravatar images when loading the image of other user's oinks in our home feed.
Things left to change:
-The information is still displaying the wrong way in the profile. That has to be changed.
Things left to implement:
-The list of followers
-The list of people I follow
-Retweets
-Oink 140-character limit