ADD FOLLOW AND UNFOLLOW FUNCTIONALITY
We added a button to the user's profile that made it possible for them to follow and unfollow other users. The button will not show up for the current user as he cannot follow or unfollow himself. The unfollow button only shows up on profiles of the people you follow and the follow button only shows up on the profiles of the people you don't follow.